refactor: 구입 금액 입력받기 리팩토링
- numberFactory에 기능이 집약적으로 있는 것은 아니지만, 추후 확장가능 용이하게 책임을 최대한 분리하였다.
