+++ src/main/java/lotto/util/factory/NumberFactory.java
@@ -1,13 +1,13 @@
 package lotto.util.factory;
 
 import camp.nextstep.edu.missionutils.Console;
+import lotto.util.ErrorMessage;
 import lotto.util.parser.NumberParser;
-import lotto.util.validator.ErrorMessage;
+import lotto.view.OutputView;
 
 public class NumberFactory {
-    private final NumberParser numberParser;
-    private static final String MINUS = "-";
     private static final String REGEXP_PATTERN_NUMBER = "^-?[\\d]*$";
+    private final NumberParser numberParser;
 
     public NumberFactory() {
         this.numberParser = new NumberParser();
@@ -18,28 +18,27 @@ public class NumberFactory {
             String input = Console.readLine();
             try {
                 validate(input);
-                int number = Integer.parseInt(input);
-                validate(number);
-                return number;
+                return numberParser.intOf(input);
             } catch (IllegalArgumentException e) {
-                System.out.println(e.getMessage());
+                OutputView.error(e);
             }
         }
     }
 
     private void validate(String input) {
-        if (input.isBlank()) {
-            throw new IllegalArgumentException(ErrorMessage.WHITESPACE_STRING);
-        }
+        validateNotBlank(input);
+        validateNumberType(input);
+    }
+
+    private static void validateNumberType(String input) {
         if (!input.matches(REGEXP_PATTERN_NUMBER)) {
             throw new IllegalArgumentException(ErrorMessage.ONLY_NUMBER);
         }
     }
 
-    private void validate(int number) {
-        if (number <= 0) {
-            throw new IllegalArgumentException(ErrorMessage.ONLY_POSITIVE_NUMBER);
+    private static void validateNotBlank(String input) {
+        if (input.isBlank()) {
+            throw new IllegalArgumentException(ErrorMessage.WHITESPACE_STRING);
         }
     }
-
 }
