--- src/main/java/lotto/util/parser/Parser.java
@@ -1,5 +1,0 @@
-package lotto.util.parser;
-
-public interface Parser {
-    public int intOf(String input);
-}
