+++ src/main/java/lotto/util/validator/NumberValidator.java
@@ -1,0 +1,11 @@
+package lotto.util.validator;
+
+import lotto.util.ErrorMessage;
+
+public class NumberValidator {
+    public void validate(int number) {
+        if (number <= 0) {
+            throw new IllegalArgumentException(ErrorMessage.ONLY_POSITIVE_NUMBER);
+        }
+    }
+}
