+++ src/main/java/lotto/view/InputView.java
@@ -3,7 +3,7 @@ package lotto.view;
 import camp.nextstep.edu.missionutils.Console;
 
 public class InputView {
-    private static final String MONEY_MESSAGE = "구입금액을 입력해 주세요."
+    private static final String MONEY_MESSAGE = "구입금액을 입력해 주세요.";
 
     public static void money() {
         System.out.println(MONEY_MESSAGE);
