+++ src/main/java/lotto/view/OutputView.java
@@ -1,0 +1,7 @@
+package lotto.view;
+
+public class OutputView {
+    public static void error(Exception e) {
+        System.out.println(e.getMessage());
+    }
+}
